Move OfflineSourceMetadata into its own file since it's used all over the project now
